+++ test/tokenize-arg-string.js
@@ -3,6 +3,7 @@
 var tokenizeArgString = require('../lib/tokenize-arg-string')
 
 require('chai').should()
+var expect = require('chai').expect
 
 describe('TokenizeArgString', function () {
   it('handles unquoted string', function () {
@@ -36,5 +37,14 @@ describe('TokenizeArgString', function () {
     args[0].should.equal('--foo')
     args[1].should.equal('hello \'world\'')
     args[2].should.equal('--bar=foo "bar"')
+  })
+
+  // https://github.com/yargs/yargs-parser/pull/100
+  // https://github.com/yargs/yargs-parser/pull/106
+  it('ignores unneeded spaces', function () {
+    var args = tokenizeArgString('  foo  bar  "foo  bar"  ')
+    args[0].should.equal('foo')
+    expect(args[1]).equal('bar')
+    expect(args[2]).equal('foo  bar')
   })
 })